Build out-of-tree successfully
If you're building out-of-tree, the generated C code
corresponding to the XML in interfaces/ ought to be in the builddir;
so it needs to be #include'd from there.

However, if you include the generated source in tarballs,
that satisfies the "make" dependency and the files won't
necessarily be rebuilt, resulting in the build failing.
To avoid that, don't distribute generated source.

Signed-off-by: Simon McVittie <simon.mcvittie@collabora.co.uk>
Bug: https://bugzilla.gnome.org/show_bug.cgi?id=695280
Reviewed-by: Ross Burton <ross@burtonini.com>